FlaskApp/FlaskApp/requirements.txt
ConfigArgParse==0.11.0
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.0
PyICU==1.8
Werkzeug==0.14.1
acme==0.10.2
argparse==1.2.1
boto==2.34.0
certbot==0.10.2
certbot-apache==0.10.2
chardet==2.3.0
click==6.7
colorama==0.3.2
configobj==5.0.6
cryptography==1.7.1
dnspython==1.15.0
enum34==1.0.3
funcsigs==1.0.2
html5lib==0.999
idna==2.0
ipaddress==1.0.16
itsdangerous==0.24
mock==2.0.0
ndg-httpsclient==0.4.2
parsedatetime==1.4
pbr==1.8.0
psutil==4.1.0
pyOpenSSL==16.0.0
pyRFC3339==1.0
pyasn1==0.1.9
python-augeas==0.5.0
pytz==2012rc0
requests==2.4.3
six==1.10.0
urllib3==1.9.1
virtualenv==16.0.0
wheel==0.24.0
wsgiref==0.1.2
zope.component==4.2.1
zope.event==4.0.3
zope.interface==4.1.1
